requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
rdkit-pypi>=2022.9.5
torch>=2.2
torch-geometric>=2.5
mlflow>=2.10
streamlit>=1.34
joblib>=1.3

